minor refactor, comments
