+++ app/Dockerfile
@@ -1,3 +1,7 @@
+# Dockerfile
+
+# Ez a file tartalmaz instrukciókat arra hogyan építsük fel a program Docker konténerét
+
 FROM python:3.10
 
 WORKDIR /app
